+++ packages/benchmark-js/src/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './StagePrefix';
+export * from './math';